Add Solution2418 class with sorting algorithms
This commit introduces the Solution2418 class in the 'leetcode.daily_challenges.easy_challenges' package. The class contains various sorting methods (such as quicksort and insertion sort) to sort people by height in reverse order. It also includes a method to swap array elements. This is a more effective approach compared to the commented sortPeople method due to improved time complexity.